feat: Creado los modulos separados, preparando para refactorizar
