+++ uploads/route.py
@@ -1,8 +1,12 @@
 from fastapi import APIRouter
 from fastapi import HTTPException
+from fastapi import BackgroundTasks
 
 from typing import Dict
 import json
+
+from uploads.logic import ImageData
+from uploads.logic import process_image_data
 
 router = APIRouter()
 
@@ -16,3 +20,8 @@ async def method_get():
 
     except FileNotFoundError:
         raise HTTPException(status_code=404, detail="Archivo no encontrado")
+
+@router.post("/upload")
+async def upload_image(data: ImageData, background_tasks: BackgroundTasks):
+    result = process_image_data(data)
+    return result
